Working monitor with sudo.
- Moved `monitor` from a shell function to a script, now it's callable
without mocking with the shell. Also, added an alias for `monitor` to
expand aliases after it.
- `sudome` is now a shell script instead of a function, now it works
around the `secure_path` option by trying to find the executable in the
path before calling `sudo`. Also, added an alias for `sudome` to expand
aliases after it.
- `__apt-daily` is now a shell script, it probably could have stayed a
shell alias but the quotes failed to pass the entire command to
`/bin/sh` as it was and I couldn't be arsed to solve it.
- Simplified `flatpak-daily`.
- `dd` shell alias is simpler and now the invocation `sudome dd` works
as expected (which was the whole point to begin with).