Add docker remote sub-command.
Add the docker remote sub-command (actually the _docker-remote script
but can be invoked as a sub-command with the wrapper script). It creates
a local socket that's forwarded to the dockerd socket on the remote host
via ssh. Requires OpenSSH 6.7 or later (for Unix socket forwarding and
%C token. Typical invocation would be `$(docker remote
host.domain.example)`. A blog post will follow.