Ansible config changes.
Per
https://docs.ansible.com/ansible/latest/reference_appendices/config.html#the-configuration-file
the ~/.ansible.cfg file is ignored if another is found first. So use
environment variables instead so they work with all of my setups (the
internded purpose). Also, for the ansible-local and
ansible-local-playbook aliases, use the Python3 interpreter that was
used to run Ansible in the first so that modules (like postgresql_privs)
that require Python module have access to them.