Specify the Redis history file with an environment variable.
The Redis CLI history file was moved to the Documentst folder so it
would synced using Nextcloud. A symlink was left in its place to point
to the new location. However it's a dangling link until the Redis CLI is
first run, which upsets the pre-commit broken link check. Instead, set
the REDISCLI_HISTFILE environment variable to the new location and
remove the symlink.