Remove the .poca symlink.
The symlink was added to keep the .poca directory in sync using
Nextcloud by having it reside inside the Documents directory which is
synced. But the subscription list is committed to the poquita repo and
the database is in the Kodi host. The directory is empty, nothing to
sync, everything can be deleted. Now the broken links pre-commit check
will be quiet.