Keybase management script.
I don't like Keybase. I don't like installing Debian packages without a
repository, I don't like the Cron jobs it adds, I don't like the fuse
mount, I don't like the processes that keep running when the I close the
application. All of this just for a mediocre password manager. But this
is not the first time I had to work with it so being as stubborn as I am
I went a different route. The workbench image has just the CLI and fuse
driver. This script start the daemon (why the hell does a password
manager needs to have a daemon running?) and mounts the Keybase
filesystem, stops and unmounts and copies the files locally and creates
some links so that scripts that have hard-coded paths would work. This
may be going way too deep in to the unsupported setup route. At least if
this doesn't work I'll have someone to blame.