Don't hang on sending notification to Nextcloud.
I tried calling curl with & at the end, but that caused output from the
notify script to spam the prompt (process XXXX has finished). Instead I
wrote a short Python script to nicely fork and silently send the message
in the background, quickly returning to the calling process (the Bash
prompt). For reference, the problem I'm trying to solve is that on
remote systems, sending the notifications and has curl waiting for the
response. This causes the prompt to hang for a few seconds which I find
irritating.