Include <workspace>/*.tfvars files with tf.
The tf wrapper adds found tfvars files under a directory with the
workspace name. So you can have multiple files with nicer separation
inside a directory if you need it (also symlinks so you can keep things
DRY). I'm a little conflicted on this change because it makes the script
more complex and I like pointing out that the wrapper is just ~20 lines
of simple shell scripting. Finally, the invocation will fail with
filenames that have spaces in them.