Planet reader.
This is something I wanted to do for a long time. Basically it's a nice
UI for me to browse some of the sites I frequent (sort of a news
reader), but geared towards a terminal. So this will just open a
terminal web browser (although a graphical browser can be used just as
well) with one of the configured sites so I don't have to type much.
There's tab completion for even less typing. Honestly, it's been on my
todo list from the days I used to carry with me my PocketCHIP and I
could have just deleted this task a long time ago. But I recently
learned about Python's webbrowser module and I wanted to try it out. I
also learned how to do tab completion in Bash so that's nice too.