wb: Interactive rcfiles update.
Updating rcfiles triggers the Git hook which in turn runs Make to update
configuration files which runs Passhole. If not interactive (and not
TTY), Passhole falls back to using Zenity which I don't include in the
Workbench image (it pulls in a lot of X and Gnome dependencies and I
feel like it's the wrong tool for this situation). Instead, I would
prefer to have Passhole ask for password as usual, so interactive (and
TTY) it is.