Rewrite tf in Python.
There's an issue with tf import because the options (-var-file=...) must
precede the other arguments (resource name and ID). Writing it in POSIX
shell without arrays is going to be a mess. I can probably write it in
Bash but I prefer Python so here we are. I also feel it's a bit more
readable.